style: implement component-based architecture and update documentation
